refactor(slack): split MessageBuilder into specialized sub-builders
Split the monolithic MessageBuilder (1097 lines, 19+ methods) into
specialized sub-builders to address God Class risk (issue #174):

- builder_tool.go: ToolMessageBuilder (ToolUse, ToolResult)
- builder_answer.go: AnswerMessageBuilder (Answer, Error)
- builder_plan.go: PlanMessageBuilder (PlanMode, ExitPlanMode, AskUserQuestion)
- builder_interactive.go: InteractiveMessageBuilder (DangerBlock, PermissionRequest)
- builder_stats.go: StatsMessageBuilder (SessionStats, CommandProgress, CommandComplete)
- builder_system.go: SystemMessageBuilder (System, User, StepStart, StepFinish, Raw, UserMessageReceived)
- builder.go: Simplified to routing layer (171 lines)

Benefits:
- Each sub-builder 100-200 lines, easier to understand
- Independent unit testing
- Easier to extend with new message types
- Maintained backward compatibility with delegate methods
